Add settings activity
An activity was added to set preferences for the
application for instance the host address and port
of the server.

If no server was defined yet, a Toast will be displayed.
Also changed the Motion termination in the server
script from terminate() to kill().

For now the created Motion process cannot be a daemon
process. Otherwise it cannot be easily killed without
reading the pid from a file.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>