Start implementing picture transmission
For displaying the picture shot which triggered the motion the
picture file name will be inserted into the message. In addition
the picture folder will be made available, hence I have extended
the server REST API. The base URI /picture will be used for
accessing the pictures and /control for taking actions on the
server.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>