Add snapshot functionality
Until now captured images from the webcam were only delivered on motion
detection. I have defined another control address which forwards the
request to the motion HTTP server to create a snapshot.

Afterwards motion captures a picture and saves it. A symbolic link called
lastsnap.jpg is created and will always point to the latest snapshot. This
way the client always knows which URL to access for getting the current
snapshot. For the Android application I have added a button to request
such a snapshot.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>