Refactor registration procedure
The device registration was executed in the UI thread which is considered
bad practice because it can freeze the user interface. I have refactored the
registration to be executed in an AsyncTask class.

In addition I have restructured the registration invocation in the
onResume method. The device is only registered again if the preferences
have changed.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>