Replace WebView with ImageView
Instead of a WebView for displaying the camera snapshots an ImageView is
more appropriate. The WebView is more an embedded browser window where the
image view merely displays the loaded image. For this foreign code was
added which encapsulates the image downloading task in a class which also
manages the image view update through an AsyncTask.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>