Prohibit server actions without registered device
As the whole device registration lifecycle is not completely figured out
yet it sometimes happen that server actions like starting the detection
are executed without any registered device at the server. This leads to
motion detection but without that any GCM messages will ever receive a
device. Hence, all actions issued via GET requests are only performed if
there is a device.

Also there was an inappropriate Toast instantion in the ServerUtilities
class which had to be removed. For the shared preferences I have replaced
the default initialization (null) with empty strings.

Signed-of-by: Konrad Reiche <konrad.reiche@gmail.com>