Improve coupling between the server and Motion
There are still problems which prevent a seamless interoperability
between Motion and the server backend. I have created a script which is
used to issue a motion alert to the server: motion_alert.py The script
is invoked by Motion.

I have created another script, snapshot_notification.py, which is used
to tell the server backend that a new snapshot has been created. This
needs to be done because from snapshot request to the actual file
creation it takes a moment or two. This synchronization problem is
solved with an Event.

The Bottle server is now backed by paste in order to be executed
multi-threaded. When the Motion process is started the camera needs a
short warm-up phase to produce pictures with sufficient lighting. Thus,
I have added a short sleep timer.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>