Add refresh snapshot event to the ImageView
Instead of using a button, like in the previous version of the Android
application, I have added the refresh event to the ImageView itself.
This way a snapshot is generated every time the user taps on the image.

In another refactoring step I have created two methods to lock,
respectively unlock the user interface. The methods are called in the
MainActivity and the onPostExecute() methods of the asynchronous
tasks.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>