Implement task to retrieve server status
It might happen that the Android application gets shut down, but if the
server endpoint is still running the motion detection the Android
application should reflect that after it is newly started. For that
reason the server status can be retrieved.

I have added a new asynchronous task which requests the status and acts
accordingly. Also I have added an example settings file which is not on
gitignore and which should be modified by the user.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>