Start implementing landscape view
For maximizing the snapshot I want to use the landscape mode. Hence,
when the phone is rotated the image should fill the whole screen. For
that I have started to override the appropriate methods and make sure
that rotation is only enabled when the motion detection is active.

By researching about this topic I have also learned about saving the
activity state. Thus, it might not be necessary anymore to retrieve the
server status every time.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>