Swallow lastsnap.jpg file creation event
With every requested snapshot Motion updates a link from lastsnap.jpg to
the last created snapshot. This is a handy utility, but sometimes a
little bit too slow. Watchdog notices this event and until now it was
interpreted as a motion alert. Hence, I have added an if-case to swallow
this event.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>