Change file event trigger to on modified
Overriding the Watchdog method for on file created seemed to be
sufficient, but on the Raspberry Pi and with large snapshots, i.e. a
resolution of 1280x720 and with a size of around 50 KB, there is a
certain gap between file creation and last write to that file.

The result of that was the following: If the client requested a
snapshot, the server would invoke a snapshot creation. The file was
created and then the server immediately responded with a HTTP 200 OK,
but since the file was still empty the response contained no bytes and
no image could be decoded on the client side.

Hence, it was clear that the relevant method to override is on file
modified as this is invoked after the file stream is closed again. Also,
I have added an exception case for the try-clause in the server script.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>