Implement state transition Idle -> Detecting
The first two methods which were removed from the Client class dealt
with the invocation request to start or stop the detection. In the idle
controller state class, both of these cases, the start and stop request,
are processed. For that reason the asynchronous task for the detection
request is now instantiated and executed in this controller state class
as well.

The whole onPostExecute method in DetectionRequest was refactored. I
removed all calls to the activity and replaced them with updates to the
model. In the end, the model will notify the observer and the interface
can then update the interface in a way that matches the model.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>