Implement restore client state
In the process of refactoring I am looking on the actions taken when the
main activity is instantiated. There it is checked, whether a previous
model state can be restored. For this I have added another operation to
the protocol.

During that refactoring I noticed that I can generalize the model and
controller classed more then before. Hence, the point where it gets more
concrete is delayed even more. For that the generic typing had to be
improved.

For the time being I have also added a set state enumeration constants
to the protocol. Probably they will be removed again, since setting the
state of the model directly by the view should be implemented by setting
the appropriate controller state.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>