Add device registration to the controller handling
Still, the first thing the Android application is trying to do, is to
set up the device registration with the GCM service. Now I have added
handling as well to the idle controller state.

For the device registration message the identifier has to be passed as
well. For now I will use the object field of the message. I have to find
out whether it is more reasonable to use the data bundle for that task.

With this commit I have also removed the set state actions from the
protocol. Instead, I am now manipualting the model state directly from
the main activity. Even though this violates the component separations,
I believe that this far cleaner than cluttering the protocol enumeration
with actions that do not belong there or creating another enumeration
for these kind of tasks.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>