Add outbox handling for issued activity actions
Finally, I am putting the pieces together for handling the messages that
are issued for the controller's outbox. For that I have introduced a new
enumeration called Action which lists the different actions that are
processed by the activity.

For now I have roughly implemented the GCM registration procedure.
Outboxing allows me to set also primitive values, like a boolean, as
object parameter for the message.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>