Refactor asynchronous tasks to use outbox handler
I have refactored the remaining asynchronous tasks to use the outbox
handler mechanism as well. With this I noticed the creation of redundant
code, where the model is updated and a message is sent to the outbox
handler. I guess I will refactor these calls into separate methods at a
later point.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>