Overwrite GCM receiver to permit custom package
In order to move the GCM intent service class into another package but
the default application package one has to extend the GCM broadcast
receiver class and overwrite the getGCMIntentServiceClassName method to
return the custom class with the full package identifier.

This commit does that and makes the needed changes to the Android
manifest file.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>