Reinstate model and controller in onCreate
Sometimes I get the error, that the inbox handler is sending messages to
a handler on a dead thread. I can only imagine, that the previous thread
is abandoned, but its reference is kept. I am still unsure how one has
to handle this properly. For now I am reinstating the model and
controller everytime at the top of onCreate.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>