Fix on notification intent snapshot loading
There were several problems with the notification intent. I admit, it is
unfitting to use the same interface for displaying the motion detection
snapshot as well. But for the time being I want to make it work. I have
removed the on new intent method and moved the code into onCreate.

With some modifications to the controller states I was able to make it
work. For the future I would like to implement a different activity for
browsing the motion detection snapshots with an available history.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>