Improve download motion snapshot
I have improved the procedures that are run when an alert notification
is used to to start the application. The filename is used to distinguish
whether the latest or just some snapshot should be requested from the
server.

This, however, is still not satisfying. I believe that there is some
natural modeling problem here. Maybe the wait state should be no state
at all. After all, in every request the client has to wait. This has
nothing to do with the actual state. I think there are only two real
states: idle and detecting.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>