Make snapshots always accessible
For security reasons I only permitted the access of snapshots if there
was a running Motion process. It might be the case, however, that the
client has stopped the detection and has left alert notifications that
can be used to download the snapshots that triggered the alert. Hence, I
have removed that restriction.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>