Add GCM Sender ID to the preferences
Until now the GCM Sender ID was hard coded as a string. I have removed
that constant from the main activity and added another option to the
shared preferences which can be filled in the settings activity.

Also I have replaced the super constructor in GCMIntentService with a
no-args constructor. Now the sender ID is passed by the overridden
method getSenderIds.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>