Update notification icon
For Android 3.0 and later the guidelines for designing status bar icons
have changed. Now these icons are composed simply of white pixels on a
transparent backdrop, with alpha blending use for smooth edges and
internal texture where appropriate.

Instead of designing an own icon I have chosen a default icon for that.
It is simply a camera.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>