Glue restore state and download snapshot together
With this commit I have decided again to glue two protocol actions
together. The restore client state action and the download snapshot
action.

With this change I have also started to create methods in the client
controller for sending messages to the handler in order to purge the
code further and separate the responsibilities.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>