Fix illegal state exception
I have found the reason why the application sometimes crashes after
trying to stop the detection. Apparently the controller state needs to
be redefined when onResume is executed. Although this bug was already
fixed with the previous commit the commit is used to document the bug
eventually. Besides that I have added Javadoc to the onResume method.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>