Move update settings out of the main activity
The settings activity still contains only few lines of code. Until now I
was forced to check whether the settings are all set twice: in onCreate
and onResume. Everytime onCreate is invoked, onResume will be invoked as
well.

Hence, I have moved the code out and override the onPause method of the
settings activity. This way the settings are only updated after the user
went on the settings activity.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>