Purge further code
As expected I have removed the method which checks whether the network
is currently available, for instance it might not be available due to
activated airplane mode. I did not find a suitable place to integrate
it. An error message is displayed anyway.

Also again some small changes to the settings activity. Next I will
start refactoring the controller task classes.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>