Create launcher activity
One of the last features I want to implement before release on the
Android application client is a launcher activity.

This activity should serve as an entry point to launch either the main
or settings activity, and, what is more, it should display instructions
when the application is started for the first time. This way users will
not download the application and think that it can work without the
appropriate backend.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>