Implement just-once instruction activity
In order to avoid confusion about how this application works I have
implemented a launcher activity which displays instructions. The user
can either select 'Continue' or 'Exit'. If the user selects 'Continue'
then the instruction will not be shown anymore.

By using a translucent and no title bar theme the activity will not be
noticed on further application starts.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>