Enable detection stop during snapshot loading
As described in issue #3 I need the functionality to stop the motion
detection while the current snapshot is still loading. Sometimes the
mobile internet connection is just too slow when all I want is to
shutdown the system before I enter my apartment.

This was done by adding task cancellation to the detection state and
enabling the toggle button when the interface is set to the detection
mode. Hence, this commit fixes #3.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>