Document and refactor model classes.
Cleaned up and documented the model classes. The previous state field
was removed completely as it is not required anymore. For the time being
the model is purely cosmetic as it does not implement any logic and the
interface does not access it for retrieving the state. But, in order to
be flexible for changes in the future this class will of course remain.
After all this is a complete implementation of the MVC like architecture
for Android applications. Also added more Javadoc.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>