Complete Javadoc
Finally, I have added Javadoc to every part of the code I could think
of. Hence, this commit fixes #5. The Javadoc generator finishes without
warnings or errors. The documentation folder was still missing from the
gitignore file. For the future the documentation could be made available
through the gh-pages branch.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>