Use Dropbox REST API for backups
I have implemented support for the Dropbox API. If the user registers
for a Dropbox API key he/she can enter it into the settings file. Then
the server will authenticate the application with the user's account.

Newly created motion snapshots are then stored in the application
folder. This commit closes #4.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>