Add device registration action to the interface
There was a fundamental problem with detecting the state of the device
registration on the server and keeping this synchronized with the state
of the client.

Now I have added a separate action for the interface which issues a
device registration. Especially, I forgot to set the device registered
boolean back to false if this was the the case.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>