Use multiprocessing for the Dropbox backups
The approach with uploading the file to the Dropbox storage directly
during the on modified event was flawed. The server became blocked and
could not respond anymore.

Hence, I have made use of Python's multiprocessing which can launch
multiple processes to outsource heavy tasks. Also this way the Global
Interpreter Lock is side-stepped.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>